requirements.txt
customtkinter==5.2.1
pillow==10.3.0
